+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - Lucene query parsing as an external library
 - Deploying to the DockerHub
 - `--presearcher` flag to specify [Lucene Monitor Presearcher](https://lucene.apache.org/core/9_1_0/monitor/org/apache/lucene/monitor/Presearcher.html)
+- more efficient `MonitorQuerySerializer`
 
 ## v2022.05.01